pyproject.toml
[project]
name = "hint"
version = "0.1.0"
description = "HINT: Hierarchical Intervention Prediction Network"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "h5py>=3.10.0",
    "hdf5plugin>=6.0.0",
    "hydra-core>=1.3.2",
    "joblib>=1.3.0",
    "lime>=0.2.0",
    "loguru>=0.7.2",
    "numpy>=1.26.0",
    "omegaconf>=2.3.0",
    "polars>=0.20.0",
    "pyarrow>=22.0.0",
    "pydantic>=2.5.0",
    "pytest>=9.0.2",
    "rich>=13.7.0",
    "scikit-learn>=1.3.0",
    "shap>=0.44.0",
    "spacy>=3.7.0",
    "torch>=2.1.0",
    "torch-ema>=0.3",
    "transformers>=4.36.0",
    "xgboost>=2.0.0",
]

[project.scripts]
# src/hint/app/main.py 내부의 main 함수를 실행
hint = "hint.app.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/hint"]

[tool.uv]
package = true
